+++ src/main.js
@@ -9,6 +9,7 @@ export { default as Code } from './components/Code';
 export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
+export { default as ProgressBar } from './components/ProgressBar';
 export { default as Popover } from './components/Popover';
 export { default as Poptip } from './components/Poptip';
 export { default as Radio } from './components/Radio';