+++ CHANGELOG.md
@@ -6,7 +6,14 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 
 ## Unreleased
 - [Feature] Add data attribute `data-oui-component` to more components and add `.highlight-react--oui` class when present to highlight components in use. (#751)
-- [Feature] Add flex-grow--none and update flex-shrink--none helpers.
+
+## 26.2.0 - 2017-07-28
+- [Feature] Add ProgressBar component. (#758)
+- [Feature] Add flex-grow--none and update flex-shrink--none helpers. (#744)
+
+## 26.1.1 - 2017-07-27
+- [Feature] Add `yarn.lock` file to OUI to keep a more reliable dependency managament than `npm`.
+- [Patch] Upgrade `jest` framework from version `18.0.0` to `20.0.4`.
 
 ## 26.1.0 - 2017-07-25
 - [Feature] Add data attribute `data-oui-component` to all outer React components for tracking. Also add URL target to display borders in the docs/app via url `#oui-showreact`. (#751)